hoist selected node info so App and JsonExplorer have its data
